Fix bug in the contract and add slippage for profit

+++ packages/shared/src/token/crypto_token.rs
@@ -57,6 +57,6 @@ impl CryptoToken {
     }
 
     pub fn convert_to_amount(&self, value: f64) -> U256 {
-        U256::from((value * self.decimals_pow) as u64)
+        U256::from((value * self.decimals_pow) as i128)
     }
 }
